+++ src/graphql/index.ts
@@ -1,1 +1,1 @@
-export * from './apolloClient';
+export * from './createApolloClient';
